+++ Scripts/SettingsHandler.gd
@@ -8,16 +8,18 @@ static var font_size: int = 18
 static var image_default: bool = false
 # godot does not have a Set datatype. lame!
 static var opened_channels: Array = []
+static var websocket_address: String = Main.DSCR_URL
 
 static func initialize() -> void:
 	# Setting names kept in PascalCase to keep backwards compatibility
 	# also for some reason TMfDS uses PascalCase too. blegh.
 	do_formatting = SaveSystem.settings.get_or_add("DoFormatting", true)
 	image_default = SaveSystem.settings.get_or_add("ImageDefault", false)
-	truncate_message_size = SaveSystem.settings.get_or_add("TruncateMessageSize", 64)
+	truncate_message_size = SaveSystem.settings.get_or_add("TruncateMessageSize", 63)
 	preferred_callsign = SaveSystem.settings.get_or_add("PreferredCallsign", -1)
 	font_size = SaveSystem.settings.get_or_add("FontSize", 18)
 	opened_channels = SaveSystem.settings.get_or_add("OpenedChannels", []).map(func (a): return int(a))
+	websocket_address = SaveSystem.settings.get_or_add("WebsocketAddress", Main.DSCR_URL)
 
 static func save() -> void:
 	SaveSystem.save_settings()
@@ -29,3 +31,4 @@ static func export() -> void:
 	SaveSystem.settings.set("PreferredCallsign", preferred_callsign)
 	SaveSystem.settings.set("FontSize", font_size)
 	SaveSystem.settings.set("OpenedChannels", opened_channels)
+	SaveSystem.settings.set("WebsocketAddress", websocket_address)